Rayt 37 tester fix linitialisation

+++ src/Raytracer/Entity/ALight.cpp
@@ -10,18 +10,21 @@
 #include <utility>
 
 namespace Raytracer {
-    ALight::ALight(std::string type, const Component::Vector3f &position, const Component::Color &color, float intensity)
+    ALight::ALight(std::string type, const Component::Vector3f &position, const Component::Color &color, double intensity)
             : AEntity(Raytracer::CompType::LIGHT, position), _typeLight(std::move(type)), _color(color), _intensity(intensity) {}
 
-    const Component::Color& ALight::getColor() const {
+    const Component::Color& ALight::getColor() const
+    {
         return _color;
     }
 
-    float ALight::getIntensity() const {
+    double ALight::getIntensity() const
+    {
         return _intensity;
     }
 
-    const std::string &ALight::getTypeLight() const {
+    const std::string &ALight::getTypeLight() const
+    {
         return _typeLight;
     }
 }
